pyproject.toml
[project]
name = "mcp-seo"
version = "0.1.0"
description = "MCP SEO - Professional SEO Analysis & Research Tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.13.0",
    "fastmcp>=2.12.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.51.0",
    "nest-asyncio>=1.6.0",
    "requests>=2.32.0",
    "pydantic>=2.11.0",
    "typing-extensions>=4.10.0",
]
